refactore the code and apply SOLID principel
