+++ app.rb
@@ -11,6 +11,36 @@ class App
     @rentals = []
   end
 
+  def display_menu
+    puts "\nPlease choose an option by entering a number:"
+    puts '1 - List all books'
+    puts '2 - List all people'
+    puts '3 - Add a person'
+    puts '4 - Add a book'
+    puts '5 - Add a rental'
+    puts '6 - List all rentals for a given person id'
+    puts '7 - Exit'
+  end
+
+  def choose_option(option)
+    case option
+    when 1
+      list_all_books
+    when 2
+      list_all_people
+    when 3
+      create_person
+    when 4
+      add_book
+    when 5
+      add_rental
+    when 6
+      list_all_rentals
+    else
+      'Wrong number please enter a number between 1 and 7'
+    end
+  end
+
   def list_all_books
     @books.each do |book|
       puts "Title: \"#{book.title}\", Author: #{book.author}"
